pyproject.toml
[project]
name = "airflow-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow[amazon]>=3.0.4",
]
